requirements.txt
PyYAML
sounddevice
webrtcvad-wheels
numpy
soundfile
requests
# Wyoming-protokollklient (PyPI)
wyoming
